Consolidated test base classes, added logout method, updated post response
